Fixed a small issue with datatypes in submission entity and its effects on other classes... Overall, the program works perfectly from signing up, verifying an account and logging in/out til adding/listing problems and creating/retrieving submissions. - Zero

+++ src/main/java/com/zerowhisper/codeforcessmallmimic/entity/Submission.java
@@ -46,10 +46,10 @@ public class Submission {
     private String userOutput;
 
     @Column(name = "time_taken")
-    private Integer timeTaken;
+    private Double timeTaken;
 
     @Column(name = "memory_taken")
-    private Integer memoryTaken;
+    private Double memoryTaken;
 
     @PrePersist
     protected void OnCreate() {
